2018.03.07 重新生成Activity类

+++ QingXie/src/main/java/com/selfcreate/qingxie/bean/activity/Activity.java
@@ -13,9 +13,9 @@ public class Activity {
 
     private Integer status;
 
-    private Integer hours;
+    private Double hours;
 
-    private Integer hourPerTime;
+    private Double hourPerTime;
 
     private Integer needVolunteers;
 
@@ -79,19 +79,19 @@ public class Activity {
         this.status = status;
     }
 
-    public Integer getHours() {
+    public Double getHours() {
         return hours;
     }
 
-    public void setHours(Integer hours) {
+    public void setHours(Double hours) {
         this.hours = hours;
     }
 
-    public Integer getHourPerTime() {
+    public Double getHourPerTime() {
         return hourPerTime;
     }
 
-    public void setHourPerTime(Integer hourPerTime) {
+    public void setHourPerTime(Double hourPerTime) {
         this.hourPerTime = hourPerTime;
     }
